requirements.txt
# Core dependencies
requests>=2.31.0
pydantic>=2.4.0
typing-extensions>=4.8.0
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.66.0
rich>=13.6.0

# API clients
openai>=1.3.0
anthropic>=0.7.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0

# Configuration management
pyyaml>=6.0.1
jsonschema>=4.19.0

# Async support
aiohttp>=3.8.0
asyncio-throttle>=1.0.2

# Text processing
regex>=2023.8.8
nltk>=3.8.1

# Logging and monitoring
loguru>=0.7.2
prometheus-client>=0.18.0

# Statistics and analysis
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Caching
redis>=5.0.0
diskcache>=5.6.3

# Development dependencies (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.1
black>=23.9.1
isort>=5.12.0
mypy>=1.6.0
pre-commit>=3.4.0
flake8>=6.1.0

# Visualization dependencies (optional)
matplotlib>=3.7.0
seaborn>=0.13.0
plotly>=5.17.0

# Benchmarking dependencies (optional)
datasets>=2.14.0
huggingface-hub>=0.17.0